Scripts on the PC can run commands on the browser
The scripts must be named following the pattern browser.*.ahk, and be located in the ‘windows’ directory. The only command currently implemented is ‘screen’.

The package.json file had to be moved to the root directory, because Node was not able to find package 'binary-split', when required by a file outside the ‘server’ directory.